pyproject.toml
[tool.poetry]
name = "kanban-life"
version = "0.1.0"
description = ""
authors = ["Gustavo <gustavosandrindattein123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
gunicorn = "^23.0.0"
dj-database-url = "^2.3.0"
python-decouple = "^3.8"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
django-s3-folder-storage = "^0.5"
django-model-utils = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
